Create JavaLoops
We use the integers a, b and n to create the following series:
(a+2^0*b),(a+2^0*b+2^1*b), ... ,(a+2^0*b+2^1*b+...+2^(n-1)*b)
You are given q queries in the form of a, b and n. For each query, print the series corresponding to the given a, b and n values as a single line of n space-separated integers.
Input Format
The first line contains an integer, q, denoting the number of queries. 
Each line i of the q subsequent lines contains three space-separated integers describing the respective a, b and n values for that query.
Constraints
* 0<q<=500
*0<=a,b<=50
*0<=n<=15